+++ lab/test-buddyshare.sh
@@ -53,9 +53,9 @@ go build -o "$BN" ./cmd/buddynet
 sudo modprobe wireguard
 
 echo "== [$(date +%T)] identities =="
-SRVPUB=$("$BN" --key "$D/srv.key" identity)
-APUB=$("$BN" --key "$D/a.key" identity)
-BPUB=$("$BN" --key "$D/b.key" identity)
+SRVPUB=$("$BN" --key "$D/srv.key" init)
+APUB=$("$BN" --key "$D/a.key" init)
+BPUB=$("$BN" --key "$D/b.key" init)
 
 echo "== [$(date +%T)] bridge topology (ns-srv/a/b on br0 in ns-sw) =="
 sudo ip netns add ns-sw; sudo ip netns add ns-srv; sudo ip netns add ns-a; sudo ip netns add ns-b